[AWIBOF-6498] Fix log group reset method to modify only log footer (#212)
* [AWIBOF-6498] Fix log group reset method to modify only log footer

+++ test/unit-tests/journal_manager/replay/replay_log_list_mock.h
@@ -12,7 +12,7 @@ public:
     using ReplayLogList::ReplayLogList;
     MOCK_METHOD(void, AddLog, (LogHandlerInterface* log), (override));
     MOCK_METHOD(bool, IsEmpty, (), (override));
-    MOCK_METHOD(void, SetLogGroupFooter, (uint64_t seqNum, LogGroupFooter footer), (override));
+    MOCK_METHOD(void, SetLogGroupFooter, (uint32_t seqNum, LogGroupFooter footer), (override));
     MOCK_METHOD(ReplayLogGroup, PopReplayLogGroup, (), (override));
 };
